Move progress report thread start VLC->main
The VLC playback handler was starting a thread to report progress back
to alexa API. In order to simplify the playback handlers some the
creation of the thread is moved to the main player playback_callback.
Another thread is already started here when calling the alexa API so it
seems more appropriate as well. The one functional change however is
that the calls to platform.indicate_playback() will happen on the
playback_handler thread. Since there seems to be no other attempts to call
platform methods such as indicate_recording(), indicate_failure(), etc
on a separate thread this change also more inline with the rest of the code.