Add playback_padding config option, helps playback
On some platforms like CHIP the playback of audio is clipped of at the
beginning or end. The SoX handler that will be added is able to use this
option to pad the beginning and ending silence. The value refers to
seconds of silence to add to the beginning or end of audio to be played
back